% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_DF.R
\name{calculate_DF}
\alias{calculate_DF}
\title{calculate differential fusion events}
\usage{
calculate_DF(obj, group, nameT, nameC, at.least.nSample = 1)
}
\arguments{
\item{obj}{a FusionSet object}

\item{group}{set the group column}

\item{nameT}{name of treat group}

\item{nameC}{name of control group}

\item{at.least.nSample}{filter the fusion events detected in at least n samples}
}
\value{
results of differential fusion events
}
\description{
find differential fusion (DF) events using DESeq2 (based on a model using the negative binomial distribution) and MAST ()
}
\examples{
\dontrun{
df_res = calculate_DF(obj, group='group',nameT='CA',nameC='NAT')
head(df_res)
}
}
\references{
Love, M.I., Huber, W., Anders, S. Moderated estimation of fold change and dispersion for RNA-seq data with DESeq2 Genome Biology 15(12):550 (2014)
McDavid A, Finak G, Yajima M (2023). \emph{MAST: Model-based Analysis of Single Cell Transcriptomics}. doi:10.18129/B9.bioc.
}
