% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fusion_object.R
\name{create_fusion_object}
\alias{create_fusion_object}
\title{create a fusion object from raw reads count}
\usage{
create_fusion_object(input.dir, min.count = 3, info = NULL)
}
\arguments{
\item{input.dir}{directory to all final-list_candidate-fusion-genes.txt renamed to sample_name.txt (character)}

\item{min.count}{filter fusion events at least min.count spanning unique reads (numeric)}

\item{info}{sample information with sample id as row names (data frame)}
}
\value{
a FusionSet object.
}
\description{
import the FusionCatcher output to create a FusionSet object
}
\examples{
\dontrun{
input.dir = system.file("extdata","fusioncatcher_output", package = "fusionMHC")
info = data.frame(row.names = c("sample_1_CA","sample_1_NAT","sample_2_CA",
                 "sample_2_NAT","sample_3_CA","sample_3_NAT"),
                  group = rep(c('CA','NAT'),3))
obj = create_fusion_object(input.dir, info=info)
}
}
