% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_fasta.R
\name{parse_fasta}
\alias{parse_fasta}
\title{prepare the reference sequence}
\usage{
parse_fasta(fasta.file)
}
\arguments{
\item{fasta.file}{fasta file with corresponding version referring to FusionCatcher output}
}
\value{
a DNAStringSet object
}
\description{
prepare the reference sequence
}
\examples{
\dontrun{
fa.set = parse_fasta(fasta.file='path/to/GRCh38.fa')
}
}
