% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gtf.R
\name{parse_gtf}
\alias{parse_gtf}
\title{prepare the gene annotation from the gtf file}
\usage{
parse_gtf(gtf.file)
}
\arguments{
\item{gtf.file}{gtf file referring to version used in FusionCatcher}
}
\value{
a gene annotation list including exon elements
}
\description{
extract the exon elements from the gtf file
}
\examples{
\dontrun{
# downloaded from //ftp.ncbi.nlm.nih.gov/refseq/MANE/MANE_human/release_1.4/MANE.GRCh38.v1.4.refseq_genomic.gtf.gz.
# you should download the corresponding version referring to FusionCatcher output
gtf.ele.list = parse_gtf(gtf.file='path/to/MANE.GRCh38.v1.4.refseq_genomic.gtf.gz')
}
}
