% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_freq_of_binding_cores.R
\name{plot_most_freq_binding_cores}
\alias{plot_most_freq_binding_cores}
\title{plot top frequent MHC binding cores}
\usage{
plot_most_freq_binding_cores(
  obj,
  topN = 10,
  group = NULL,
  hla.type = c("HLA-A0201", "HLA-A1101"),
  point.size = 5,
  bar.color = "gray",
  bar.width = 0.1
)
}
\arguments{
\item{obj}{a FusionSet object}

\item{topN}{show top N binding cores}

\item{group}{set the group name}

\item{hla.type}{set the HLA type}

\item{point.size}{set point.size}

\item{bar.color}{set bar.color}

\item{bar.width}{set bar.width}
}
\value{
a ggplot2 object
}
\description{
barplot shows the most frequent MHC binding cores
}
\examples{
\dontrun{
top_cores = plot_most_freq_binding_cores(obj, topN=10, group="group")
}
}
