% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_occurrence.R
\name{plot_occurrence}
\alias{plot_occurrence}
\title{heatmap shows occurrence for each HLA type}
\usage{
plot_occurrence(
  obj,
  hla.type = c("HLA-A0201", "HLA-A1101", "both"),
  anno.sample = NULL,
  anno.colors = NULL,
  col = c("blue", "green", "yellow"),
  fusion.to.show = NULL,
  top.to.show = 10,
  show_column_names = T,
  show_row_names = T,
  ...
)
}
\arguments{
\item{obj}{a FusionSet object}

\item{hla.type}{set the HLA type}

\item{anno.sample}{data frame to annotate the samples}

\item{anno.colors}{a list of colors}

\item{col}{set 3 colors if hla.type is "both". otherwise 2 colors.}

\item{fusion.to.show}{set fusion names to show (chimericRNA)}

\item{top.to.show}{show n fusion events}

\item{show_column_names}{TRUE or FALSE}

\item{show_row_names}{TRUE or FALSE}

\item{...}{arguments passed to ComplexHeatmap::Heatmap}
}
\value{
a FusionSet object
}
\description{
heatmap shows occurrence for each HLA type
}
\examples{
\dontrun{
obj = plot_occurrence(obj, hla.type = 'both', col = c("blue",'green','yellow'),
                      anno.sample = obj@sample.data,
                      anno.colors = list(group = c("CA"="red","NAT"='green')),
                      top.to.show = 10)
}
}
