% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_summary.R
\name{plot_sample_summary}
\alias{plot_sample_summary}
\title{plot basic summary}
\usage{
plot_sample_summary(
  obj,
  features = c("nFusion", "nCount"),
  group = NULL,
  group.colors = NULL,
  plot.type = "barplot",
  topN = 50,
  bar.width = 0.5,
  ncol = 2
)
}
\arguments{
\item{obj}{a FusionSet object}

\item{features}{colnames in obj@sample.data}

\item{group}{specify group name}

\item{group.colors}{set group colors}

\item{plot.type}{boxplot or barplot}

\item{topN}{show top N samples}

\item{bar.width}{set bar.width}

\item{ncol}{number of columns to combine multiple feature plots to}
}
\value{
a ggplot2 object
}
\description{
plot basic summary
}
\examples{
\dontrun{
obj = plot_sample_summary(obj,features = c('nFusion','nCount'), group='group')
}
}
