% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_MHC_affinity.R
\name{predict_MHC_affinity}
\alias{predict_MHC_affinity}
\title{This function implements the prediction of binding affinity to MHC molecules for each peptide}
\usage{
predict_MHC_affinity(
  obj,
  path.to.netMHC = "path/to/netMHC-4.0/",
  temp.dir = tempdir(),
  n.cores = parallel::detectCores()/2
)
}
\arguments{
\item{obj}{a FusionSet object}

\item{path.to.netMHC}{path to netMHC-4.0}

\item{temp.dir}{specify temporary directory}

\item{n.cores}{number of parallel cores}
}
\value{
returns a \code{FusionSet} object with slot occurrence and renew slot peptide with extra columns prefix with HLA-*).
the binding details were added to obj@peptide, prefix with corresponding HLA type. occurrence matrix represents the peptides with MHC binding affinity for each HLA type in each sample.
}
\description{
predict the binding of each peptide in obj@peptide to MHC class I molecules, i.e, HLA-A0201 and HLA-A1101 using netMHC-4.0 (https://services.healthtech.dtu.dk/services/NetMHC-4.0/).
}
\examples{
\dontrun{
obj = predict_MHC_affinity(obj)
}
}
